rename: model, extensions 분리
